pyproject.toml
[tool.poetry]
name = "sharejoy_be"
version = "0.1.0"
description = ""
authors = ["Paweł Benkowski <pawlo97.pb@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.0"
djangorestframework = "^3.13.1"
psycopg2 = "^2.9.3"
python-decouple = "^3.5"
django-cors-headers = "^3.10.1"
drf-spectacular = "^0.21.1"

[tool.poetry.dev-dependencies]
mypy = "^0.930"
isort = "^5.10.1"
black = "20.8b1"

[tool.black]
line-length = 79

[tool.isort]
profile = "django"
known_django = "django"
known_rest = "rest_framework"
sections = ["FUTURE", "STDLIB", "DJANGO", "REST", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
import_heading_stdlib = "Builtins"
import_heading_django = "Django"
import_heading_firstparty = "First Party"
import_heading_thirdparty = "Third party"
import_heading_localfolder = "Local"
import_heading_rest = "Django REST Framework"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
